TR11750/postgres

+++ weekly/PRB_StrainJAX2.py
@@ -110,7 +110,8 @@ def mmrrc():
     
     # Retrieve all Strains that have a MMRRC ID and whose Alleles are used in a Genotype
 
-    db.sql('drop table #strains', None)
+    if os.environ['DB_TYPE'] == 'postgres':
+        db.sql('drop table #strains', None)
 
     db.sql('''
 	select distinct s._Strain_key,
